Start from scratch
I was creating a heap so that I could use a priority queue for the
notes that were yet to play, with the beat the note should play on
specified as the priority.

However, starting with this was complicated, and I found it hard to
think of how it would fit in to the sequencer. I’ve decided to start by
writing a basic sequencer that loops over a small, fixed number of
beats – once I’ve got this working I’ll expand and refine it from there.