Create basic sequencer
This is a very basic sequencer which forms a starting point. At the
moment, it loops a hardcoded sequence of frequencies indefinitely.

I’m using window.setInterval() to call a scheduling function, which
schedules notes that are due to happen soon.

I’m representing the beats using just the numbers 0-7, and the time for
each beat is worked out from the tempo and the time the last beat
happened at. I’d like to have a timing system that allows arbitrary
beat timing rather than fixing the notes to a grid. For a fixed tempo
this should be straightforward, but allowing the tempo to change (which
I would like to be possible) may require more effort.

At the moment the frequencies are hardcoded – I would like to have a
note system, possibly using a scale and key as well.

One thing to wonder about is whether to use MIDI for the note system
and the timing. It might make things easier, but I’m wondering whether
it might be unnecessarily complicated.

Later, it’d be nice to structure the sequencer in such a way that it
can play notes on any other synth module. For this, MIDI is definitely
going to be useful.