Correct next note time update in sequencer.play()
The next note time is an absolute time, and I was setting it to the
secondsFromNow value – if play was called when audioContext.currentTime
was greater than the value being passed to secondsFromNow, the next
note would be scheduled into the past.

From there, I think the sequencer would loop through the sequence
(scheduling notes to be played in the past) until it caught up to the
current time, meaning that resuming wouldn’t necessarily resume at the
same part of the sequence.