Write beatgrid object to handle beat–time conversion
This object will be responsible for converting between beat
representations and time representations.

I’ll need to think about whether this object should be responsible for
seeking and looping behaviour, and maintaining the current play
position. I think it would be better to separate that behaviour from
the beat–time conversion and put it in its own object.