Delegate tempo and nextNoteTime to BeatGrid object
As the beat grid is going to handle beat-time conversions, it makes
sense to keep the tempo information just with the beat grid – don’t
have to have it in multiple places.

Rather than storing the next note time, as well, the sequencer can just
store a beat number and use the beat grid to convert that into a time.

This does mean that I had to change the play function – it was relying
on the next note time property to schedule the start of playback. I’ve
changed it to schedule playback the way stop and pause do.