Rename and restructure BeatGrid to fit new concept
When I was thinking about the BeatGrid before, I was thinking that a
sequencer would have one of them, and that would define the
relationship between the beat space and the time space.

Then I realised that that link is defined by playback, in a sense, and
so it makes sense for each playing back (say each Player object) to
have its own BeatGrid. Until playback is started or scheduled, there
isn’t a concrete relationship between the beat space and the time space
(except I guess for the tempo).

I’ve also renamed the object to BeatTimeline, as I think it
communicates this concept better.

There’s still the question of whether BeatTimelines should be mutable
or not. I’m leaning towards making them immutable. This would allow
sharing them more easily between multiple players, but means that
conceptually, looping playback involves creating a new beat timeline
that’s offset from the old one, and switching to it at the right time.

That’s a bit unexpected, but I think it’s the right way to go. Leads to
interesting concepts like having different tempos for different tracks
in a multitrack sequence.

I’ve also allowed the tempo to be specified from the outside, rather
than hardcoded in the BeatTimeline class, and I’ve fixed the looping
somewhat.

The looping will needs some work, though – I want to create a copy with
a specific beat offset from the current timeline. At the moment the
copy is defined by when it’s created, which isn’t very pleasant. I want
to be able to schedule playback by creating timelines defined at
specific times/beats.