Fix looping
Before now, the looping copied the timeline and started at the time the
copy function was called, meaning it restarted slightly early, and
didn’t loop smoothly.

I’ve fixed this by enabling copies to be accurately time-shifted,
either by a specific time in the base of AudioContext.currentTime() or
by a number of beats within the timeline – so for a loop of 8 beats,
once you’ve scheduled for the last beat, you create a copy of the
timeline but shift its reference time forward by 8 beats, so the time
for the copy’s first note is now 1 beat after the original’s last note.