Remove extra prototype links for timeline copies
For timeline copies, I don’t think it makes sense to link the copies to
the original timeline – instead I think they should be linked to the
original’s prototype.

In the looping case, where I’m replacing the timeline with a copy of
itself, the prototype chain will gradually become longer and longer as
the looping continues, which couldn’t be good.

On the other hand, having the copies linked to the original would
possibly be interesting for propagating changes across related
timelines (e.g. creating a few parallel time-shifted timelines and then
changing the beat and time functions across all of them). I’ll have to
think about this one – I don’t know if it sits well with the way I’m
using them at the moment, but it may be a better approach.