Add play and pause buttons
To start off the user interface, I’m adding play and pause buttons,
which will start and stop the sequencer by calling its play() and
pause() methods.

At the moment, the underlying functions don’t function as I would like –
the pause function should halt playback at the current position in the
sequence, and the play function should resume from that point when it’s
pressed. These need to be updated to fit with the new timeline object.