Don’t pass audio context to timeline when initialising
I changed the timeline object before now to take a reference time
(along with a tempo and a reference beat) rather than an audio context.
I think limiting how many things have a reference to the audio context
will make this code more reusable in other contexts.

However, I forgot to change how the timeline is initialised when the
sequencer is created – I’m now passing it the current time instead of
the audio context, which it was ignoring anyway.

The timeline not having a reference to the audio context means that it
has no internal concept of the current time, but I think that makes
sense anyway, keeping the current time with the scheduling code.