Create a new timeline when resuming playback
Before now, when playback resumed, it acted as if the output was muted,
as the timeline was staying the same – if you pause for 2 beats and
then resume, you will resume 2 beats later than you paused.

To solve this, I’m creating a new timeline when playback is started
again, using the current time and the next beat as its reference.