Add slider to control playback tempo
The implementation of this means that a new timeline object is created
every time the tempo is changed, and since that’s using the input event
of the slider, every increment of the slider creates a new object. I’m
not sure if that’s too much load and is an argument for making the
timeline objects mutable – for the moment I’m going to continue having
them immutable and see how things go.

Another thing that’s not quite satisfactory is that the tempo changes
take effect from the next beat rather than the current one, which is
quite noticeable when we’re starting from slow tempos.

I don’t know if this is definitely something that needs changing – it
might be that allowing greater division of notes makes this less
annoying.