Convert code to ES6 module structure
This organisation of code will be better, separating the creation of the
sequencer components and so on from their usage in the example.

However, at the moment this doesn’t appear to be working – breakpoints
in the index file aren’t hit, and nothing happens when you press the
buttons (though you also get no errors). I’ll need to figure out what
the problem is here.