Create basic note and note-sequence object
I haven’t incorporated the objects into the current sequencer yet, but
it should be possible to do that pretty easily. For the moment the note
sequence object maintains a sorted array of the notes and does a linear
search over them when adding notes or getting notes.

A good next implementation might be a bucket list, or possibly a skip
list – the most important operation is playback, I think, so getting
notes out should be the priority.