Fix list sorting in BasicNoteSequence
When finding the position of a note, if it didn’t find a position, it
used the length of the list - 1 as the position. This meant that if the
note was meant to go at the end of the list, it would be put in
second-last place instead when it was added. It also meant that the
last note in the list was never played.