Move notes according to the drag feedback image
The notes were being moved just according to the pointer position, so if
you grabbed a note in its middle and dropped it somewhere, it would
reappear further right of where you expected it to.

Now the code detects the offset of the mouse so that the block will
reappear exactly where the drag feedback image was when you dropped.