Clip note positions onscreen
This makes it much more useful to move notes around.

However, the whole setup is overworked – rather than creating a new class here, I should be able to just pass a new function into the scaler which is rounded. The way this is now is awkward and inefficient, because it iterates over the provided values twice.